a8n: Rename enterprise/pkg to enterprise/internal (#6855)
