+++ enterprise/cmd/frontend/internal/dotcom/billing/plans.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/licensing"
-	"github.com/sourcegraph/sourcegraph/enterprise/pkg/license"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/license"
 	stripe "github.com/stripe/stripe-go"
 	"github.com/stripe/stripe-go/plan"
 )
