+++ enterprise/cmd/frontend/internal/licensing/enforcement_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/pkg/license"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/license"
 )
 
 func TestEnforcementPreCreateUser(t *testing.T) {
