+++ enterprise/cmd/frontend/internal/licensing/features_test.go
@@ -3,7 +3,7 @@ package licensing
 import (
 	"testing"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/pkg/license"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/license"
 )
 
 func TestIsFeatureEnabled(t *testing.T) {
