+++ enterprise/cmd/frontend/main.go
@@ -23,10 +23,10 @@ import (
 	_ "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/graphqlbackend"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/licensing"
 	_ "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/registry"
-	"github.com/sourcegraph/sourcegraph/enterprise/pkg/a8n"
-	a8nResolvers "github.com/sourcegraph/sourcegraph/enterprise/pkg/a8n/resolvers"
-	"github.com/sourcegraph/sourcegraph/enterprise/pkg/codeintel/lsifserver/proxy"
-	codeIntelResolvers "github.com/sourcegraph/sourcegraph/enterprise/pkg/codeintel/resolvers"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/a8n"
+	a8nResolvers "github.com/sourcegraph/sourcegraph/enterprise/internal/a8n/resolvers"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/lsifserver/proxy"
+	codeIntelResolvers "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/resolvers"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
 )
