+++ enterprise/internal/codeintel/lsifserver/proxy/auth.go
