+++ enterprise/internal/license/generate-license.go
@@ -26,7 +26,7 @@ import (
 	"log"
 	"time"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/pkg/license"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/license"
 	"golang.org/x/crypto/ssh"
 )
