+++ enterprise/internal/license/license.go
