+++ enterprise/internal/license/license_test.go
